scripts/pyproject.toml
[project]
name = "claude-howto"
version = "1.0.0"
description = "Claude Code 实用教程及 EPUB 构建工具"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "ebooklib",
    "markdown",
    "beautifulsoup4",
    "httpx",
    "pillow",
    "tenacity",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.pytest.ini_options]
testpaths = ["scripts/tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["scripts"]
include = ["build_epub"]
exclude = ["tests", "tests.*"]

# =============================================================================
# Ruff 配置
# =============================================================================
[tool.ruff]
target-version = "py310"
line-length = 88
include = ["scripts/**/*.py"]
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    ".pytest_cache",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle 错误
    "W",      # pycodestyle 警告
    "F",      # Pyflakes
    "I",      # isort（导入排序）
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # Ruff 专属规则
    "PTH",    # flake8-use-pathlib
    "PL",     # Pylint
    "PERF",   # Perflint
]
ignore = [
    "E501",     # 行过长（由格式化器处理）
    "PLR0913",  # 参数过多
    "PLR2004",  # 魔法值比较
    "PLR0915",  # 语句过多
    "PERF203",  # 循环内 try-except（错误处理场景可接受）
    "PERF403",  # 字典推导式（可读性偏好）
    "TC003",    # 类型检查导入（非关键）
    "PLC0415",  # 非顶层导入（惰性导入可接受）
    "RUF005",   # 集合拼接（可读性偏好）
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["build_epub"]
force-single-line = false
combine-as-imports = true

[tool.ruff.lint.per-file-ignores]
"scripts/tests/*.py" = ["S101", "PLR2004"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
docstring-code-format = true

# =============================================================================
# Bandit 配置
# =============================================================================
[tool.bandit]
targets = ["scripts"]
exclude_dirs = ["scripts/tests", ".venv", "__pycache__"]
skips = ["B101", "B113"]  # B113：httpx 超时误报（已设置 timeout）
